Menambah Session Pada PostController

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+    private function currentUserId()
+    {
+        return session('current_user_id');
+    }
+
     public function index(Request $request)
     {
         $search = $request->search;
@@ -23,18 +28,25 @@ class PostController extends Controller
 
     public function create()
     {
-        return view ('posts.create');
+        if(!$this->currentUserId()) {
+            return redirect('/login') -> with('Error', 'Tolong Login Terlebih Dahulu!');
+        }
+        return view('posts.create'); 
     }
 
     public function store(Request $request)
     {
+        if(!$this->currentUserId()) {
+            return redirect('/login') -> with('Error', 'Tolong Login Terlebih Dahulu!');
+        }
+
         $request->validate([
             'caption' => 'required|string',
             'media' => 'nullable|string',
         ]);
 
         Post::create([
-            'user_id' => 1,
+            'user_id' => $this->currentUserId(),
             'caption' => $request->caption,
             'media' => $request->media,
         ]);
@@ -51,11 +63,19 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        if($post->user_id !== $this->currentUserId()) {
+            return redirect() -> route('posts.index') -> with ('Error', 'Anda Tidak Punya Akses Untuk Mengedit Postingan Ini!');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
     public function update(Request $request, Post $post)
     {
+        if($post->user_id !== $this->currentUserId()) {
+            return redirect() -> route('posts.index') -> with ('Error', 'Anda Tidak Punya Akses Untuk Mengupdate Postingan Ini!');
+        }
+
         $request->validate([
             'caption' => 'required|string',
             'media' => 'nullable|string',
@@ -68,6 +88,10 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        if($post->user_id !== $this->currentUserId()) {
+            return redirect() -> route('posts.index') -> with ('Error', 'Anda Tidak Punya Akses Untuk Menghapus Postingan Ini!');
+        }
+
         $post->delete();
 
         return redirect()->route('posts.index')->with('success', "Post Deleted Successfully");
